shinesolutions/aem-helloworld-author-dispatcher #5 (#6)
* shinesolutions/aem-helloworld-author-dispatcher #5
* CPU usage in the dispatchers was too high and instances keep terminating before scale-up
Fix: Turn on keep-alives to reduce CPU utilization and improve response time. With keep-alives on, the load balancer doesn't need to establish a new TCP connection for every HTTP request.
KeepAlive is not set to On to fix the CPU usage problem.

* ALB terminates the connections after 60 seconds. so apache configuration should be higher than it.
Fix: When the keep-alive option is enabled, choose a longer keep-alive timeout than the load balancer idle timeout.
keep-alive time out is not set to 65. which is 5 seconds longer than default ALB timeout.

* Maximum number of keep-alive requests was low, so the queue will be full and and connection will time out with 5xx error
Fix: This option sets how many requests a single TCP connection serves when keep-alive is on. For optimal usage of resources, set the maximum number of keep-alive requests to 100 or higher.
This value is now set to 0 which means there wont be any limitation in number of keep-alive connections.

* Clarify Apache config change.

Co-authored-by: Cliffano Subagio <cliffano@gmail.com>